relocates excludeOccupiedSquares function to be reusable
